+++ test/zSchemaCoverage.test.js
@@ -15,6 +15,7 @@ const {
 } = require('mdast-builder');
 const unified = require('unified');
 const stringify = require('remark-stringify');
+const gfm = require('remark-gfm');
 const fs = require('fs-extra');
 const { report } = require('../lib/keywords');
 /* eslint-env mocha */
@@ -131,15 +132,15 @@ after('Generating Schema Coverage Report', () => {
     .values(allkeywords)
     .reduce((p, v) => [...p, ...v], [])));
   const allkeywordssupported = allkeywordsplain
-    .filter(keyword => report().has(keyword));
+    .filter((keyword) => report().has(keyword));
 
-  const overall = Math.floor(100 * allkeywordssupported.length / allkeywordsplain.length);
+  const overall = Math.floor((100 * allkeywordssupported.length) / allkeywordsplain.length);
 
   const sections = Object.entries(allkeywords).map(([name, keywords]) => {
     const [label, url] = name.split(', ');
     const coverage = Math.floor(
-      100
-      * keywords.filter(keyword => report().has(keyword)).length
+      (100
+      * keywords.filter((keyword) => report().has(keyword)).length)
       / keywords.length,
     );
     return [
@@ -150,7 +151,7 @@ after('Generating Schema Coverage Report', () => {
           tableCell(text('Keyword')),
           tableCell(text('Supported')),
         ]),
-        ...keywords.sort().map(keyword => tableRow([
+        ...keywords.sort().map((keyword) => tableRow([
           tableCell(inlineCode(keyword)),
           tableCell(text(report().has(keyword) ? 'Yes' : 'No')),
         ])),
@@ -165,6 +166,7 @@ after('Generating Schema Coverage Report', () => {
   ]);
 
   const processor = unified()
+    .use(gfm)
     .use(stringify);
 
   const output = processor.stringify(mdast);